eth, internal/web3ext: implement debug_storageRangeAt
storageRangeAt allows 'paging' through the storage of an account. The
complexity of this is superlinear because the trie contains hashed
storage keys. The preimage of certain keys may be missing if the node
was fast-synced. The API returns both the preimage and the hashed key.

Transaction pre-state recomputation logic is shared with tracing.
ForEachStorage had to be adapted in order to expose the underlying trie
keys.